edit step3 code
