+++ src/main/java/Main.java
@@ -8,12 +8,13 @@ public class Main {
         List<String> points = Input.pointInput(pointSet);
         Line line = new Line(points);
         Draw draw = new Draw(points);
+        Square square = new Square(points);
         draw.drawline();
         System.out.print(String.format("%3s","+"));
         System.out.println("------------------------------------------------");
         draw.printNumber();
         System.out.println();
-        ResultView.printResult(line);
+        ResultView.printResult(square);
 
     }
 }
